Assistant checkpoint: Update Twitter authentication with OAuth2
Assistant generated file changes:
- src/twitter_poster.py: Update Twitter authentication implementation

---

User prompt:

This issue was solved for me by a Contributor to the Tweepy project, Harmon758, via discussion # 2006 on the Tweepy GitHub.

The solution required two things. First, setting the user_auth parameter to False when calling the .get_me() method. This must be done because the function defaults to OAuth 1.0 if not.

And then second, because after I did this I got a 401 unauthorized error, I needed to tweak the format of the Access Token that I was passing through to the Client object. I was trying to pass through ‘access_token’ in the above code, but it should have been ‘access_token[“access_token”]’. The return from the .fetch_token() method is a dictionary that contains the required field.

Correct code:

Build handler with:

oauth2_user_handler = tweepy.OAuth2UserHandler(
    client_id='your_client_id',
    redirect_uri="your_redirect_uri",
    scope=["tweet.read", "and any others you need"],
    # Client Secret is only necessary if using a confidential client
    client_secret='your_client_secret'
    )
Redirect to Twitter for authorization with:

def auth(request):

    return HttpResponseRedirect(oauth2_user_handler.get_authorization_url())
And finally, callback to:

def callback(request):

    access_token = oauth2_user_handler.fetch_token(request.build_absolute_uri())

    user_client = tweepy.Client(
        access_token["access_token"]
    )

    me = user_client.get_me(user_auth=False)

    return HttpResponse()

Refer this solution to debug the issue